chore: refactor application folder structure
- moved folders into a server parent folder

- moved files into appropriate folders

[Finishes #161412030]
