+++ .gitignore
@@ -33,4 +33,3 @@ node_modules
 
 # Optional REPL history
 .node_repl_history
-
