+++ server/middleware/errorHandler.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
 
-router.use("/", require("./users"));
+router.use("/", require("../routes/users"));
 
 router.use(function(err, req, res, next) {
     if (err.name === "ValidationError") {
